requirements.txt
Django==1.10.5
gunicorn==19.4.1
celery==4.0.0
raven==5.24.0
dj-database-url==0.3.0
requests==2.9.1
djangorestframework==3.5.4
jsonfield==1.0.0
structlog==16.1.0
python-dateutil==2.4.1
psycopg2==2.5.3
django-filter==0.12.0
django-debug-toolbar==1.2.1
django-fsm==2.3.0
xlwt==1.0.0
django-phonenumber-field==1.1.0
xlrd==0.9.3


